requirements.txt
# Core compatibility versions (IMPORTANT - stable with Python 3.11)
numpy==1.26.4
pydantic==1.10.13

# Web UI
streamlit

# LangChain stack (stable working combo)
langchain==0.2.16
langchain-core==0.2.40
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-huggingface

# Vector store
chromadb==0.4.24

# Embeddings
sentence-transformers

# News API
newsapi-python
newsapi

# Environment management
python-dotenv

# HTTP requests (used by NewsAPI internally)
requests